updated validation schema to reflect device schema

+++ routes/devices/devices-route.js
@@ -6,7 +6,7 @@ const {
   deviceValidation,
   idValidation,
 } = require('../../middlewares/validation.js');
-const authorized = require('../../middlewares/Authorized.js');
+// const authorized = require('../../middlewares/Authorized.js');
 
 // /api/devices
 
@@ -37,10 +37,12 @@ router.post('/', restricted, deviceValidation, async (req, res) => {
  */
 router.get('/', restricted, async (req, res) => {
   try {
-    const { userID } = await userDB.findByUUID(req.headers.decodedToken.uid);
-    if (userID) {
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+    if (foundUser && foundUser.is_admin) {
       const devices = await db.findAll();
       res.status(200).json(devices);
+    } else {
+      res.status(403).json({ message: 'Access denied!' });
     }
   } catch ({ message }) {
     res.status(500).json({ message: 'Unable to retrieve devices.' });
@@ -54,9 +56,9 @@ router.get('/', restricted, async (req, res) => {
 router.get('/user-devices', restricted, async (req, res) => {
   const userUID = req.headers.decodedToken.uid;
   try {
-    const { userID } = await userDB.findByUUID(userUID);
-    if (userID) {
-      const userDevices = await db.findALLByUserID(userID);
+    const foundUser = await userDB.findByUUID(userUID);
+    if (foundUser) {
+      const userDevices = await db.findALLByUserID(foundUser.id);
       res.status(200).json(userDevices);
     } else {
       res.status(404).json({ message: `Devices not found!` });
@@ -73,9 +75,12 @@ router.get('/user-devices', restricted, async (req, res) => {
 router.get('/user-devices/:id', restricted, async (req, res) => {
   const userUID = req.headers.decodedToken.uid;
   try {
-    const { userID } = await userDB.findByUUID(userUID);
-    if (userID) {
-      const userDevice = await db.findByIdAndUserID(userID, req.params.id);
+    const foundUser = await userDB.findByUUID(userUID);
+    if (foundUser) {
+      const userDevice = await db.findByIdAndUserID(
+        foundUser.id,
+        req.params.id
+      );
       res.status(200).json(userDevice);
     } else {
       res.status(404).json({ message: `Device not found!` });
@@ -92,8 +97,9 @@ router.get('/user-devices/:id', restricted, async (req, res) => {
  */
 router.get('/:id', restricted, idValidation, async (req, res) => {
   try {
-    const device = await db.findById(req.id);
-    if (device) {
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+    if (foundUser && foundUser.is_admin) {
+      const device = await db.findById(req.id);
       res.status(200).json(device);
     } else {
       res.status(404).json({ message: `Device not found!` });
@@ -115,12 +121,12 @@ router.put(
   deviceValidation,
   async (req, res) => {
     try {
-      const device = await db.findById(req.id);
-      if (!device) {
-        res.status(404).json({ message: 'Device not found.' });
+      const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+      if (foundUser && foundUser.is_admin) {
+        const updatedDevice = await db.update(req.id, req.update);
+        res.status(201).json(updatedDevice);
       } else {
-        const device = await db.update(req.id, req.update);
-        res.status(201).json(device);
+        res.status(403).json({ message: 'Access denied!' });
       }
     } catch ({ message }) {
       res.status(500).json({ message: 'Unable to update the device.' });
@@ -135,9 +141,17 @@ router.put(
  */
 router.delete('/:id', restricted, idValidation, async (req, res) => {
   try {
-    const device = await db.remove(req.id);
-    if (!device) res.status(404).json({ message: 'Device not found' });
-    else res.status(200).json({ message: 'Device successfully deleted.' });
+    const foundUser = await userDB.findByUUID(req.headers.decodedToken.uid);
+    if (foundUser && foundUser.is_admin) {
+      const device = await db.remove(req.id);
+      if (device) {
+        res.status(200).json({ message: 'Device successfully deleted.' });
+      } else {
+        res.status(404).json({ message: 'Device not found' });
+      }
+    } else {
+      res.status(403).json({ message: 'Access denied!' });
+    }
   } catch ({ message }) {
     res.status(500).json({ message: 'Unable to delete the device.' });
   }
